+++ kasa/smartdevice.py
@@ -19,6 +19,7 @@ from datetime import datetime, timedelta
 from enum import Enum, auto
 from typing import Any, Dict, List, Optional
 
+from .emeterstatus import EmeterStatus
 from .exceptions import SmartDeviceException
 from .protocol import TPLinkSmartHomeProtocol
 
@@ -48,48 +49,6 @@ class WifiNetwork:
     bssid: Optional[str] = None
     channel: Optional[int] = None
     rssi: Optional[int] = None
-
-
-class EmeterStatus(dict):
-    """Container for converting different representations of emeter data.
-
-    Newer FW/HW versions postfix the variable names with the used units,
-    where-as the olders do not have this feature.
-
-    This class automatically converts between these two to allow
-    backwards and forwards compatibility.
-    """
-
-    def __getitem__(self, item):
-        valid_keys = [
-            "voltage_mv",
-            "power_mw",
-            "current_ma",
-            "energy_wh",
-            "total_wh",
-            "voltage",
-            "power",
-            "current",
-            "total",
-            "energy",
-        ]
-
-        # 1. if requested data is available, return it
-        if item in super().keys():
-            return super().__getitem__(item)
-        # otherwise decide how to convert it
-        else:
-            if item not in valid_keys:
-                raise KeyError(item)
-            if "_" in item:  # upscale
-                return super().__getitem__(item[: item.find("_")]) * 1000
-            else:  # downscale
-                for i in super().keys():
-                    if i.startswith(item):
-                        return self.__getitem__(i) / 1000
-
-                _LOGGER.debug(f"Unable to find value for '{item}'")
-                return None
 
 
 def requires_update(f):
@@ -202,7 +161,7 @@ class SmartDevice:
         >>> dev.has_emeter
         True
         >>> dev.emeter_realtime
-        {'current': 0.015342, 'err_code': 0, 'power': 0.983971, 'total': 32.448, 'voltage': 235.595234}
+        <EmeterStatus power=0.983971 voltage=235.595234 current=0.015342 total=32.448>
         >>> dev.emeter_today
         >>> dev.emeter_this_month
